heated/requirements.txt
aiohttp
asyncio
discord.py
psutil
nextcord
aiodns
humanfriendly
orjson
psutil
typing_extensions
jishaku